update to totalmilliseconds. loading bar in some cases wouldn't load and the fade wouldn't fade in. because milliseconds was returning something too small or 0;

+++ Core/Menu/LGSG/Module_main_menu_LGSG.cs
@@ -61,7 +61,7 @@ namespace OpenVIII
             }
         }
 
-        private static float PercentLoaded { get; set; } = .5f;
+        private static double PercentLoaded { get; set; } = .5f;
         public static Vector2 TextScale { get; } = new Vector2(2.545455f, 3.0375f);
 
         #endregion Properties
